+++ pkg/deployment/resources/internal_exporter.go
@@ -1,5 +1,5 @@
 //
-// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2024 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -50,7 +50,7 @@ func ArangodbInternalExporterContainer(image string, args []string, livenessProb
 		Command: append([]string{exePath, "exporter"}, args...),
 		Ports: []core.ContainerPort{
 			{
-				Name:          "exporter",
+				Name:          shared.ExporterContainerName,
 				ContainerPort: int32(port),
 				Protocol:      core.ProtocolTCP,
 			},