+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - (Bugfix) Fix Image Error Propagation
 - (Feature) JobScheduler Coverage
 - (Feature) JobScheduler Volumes, Probes, Lifecycle and Ports integration
+- (Feature) Merge ArangoDB Usage Metrics
 
 ## [1.2.38](https://github.com/arangodb/kube-arangodb/tree/1.2.38) (2024-02-22)
 - (Feature) Extract GRPC Server